Implement array expression parsing in the `parser` package
- Introduce `parseArrayExpr` function to handle array expressions, allowing for the parsing of elements within square brackets.
- Enhance the `parsePrimaryExpr` function to recognize array expressions and delegate parsing to `parseArrayExpr`.
- Add support for empty arrays and improve error handling for missing closing brackets, providing diagnostic suggestions for corrections.
- Update the `Exprs` structure in the `ast` package to include an arena for array expressions, facilitating efficient memory management.

These changes significantly enhance the parser's capabilities to interpret array expressions, improving the overall syntax handling in the Surge language.
