+++ internal/parser/let.go
@@ -37,32 +37,33 @@ func (p *Parser) parseLetBinding() (LetBinding, bool) {
 	if p.at(token.Assign) {
 		tokAssign := p.advance() // съедаем '='
 		var ok bool
+		beforeErrors := p.opts.CurrentErrors
 		valueID, ok = p.parseExpr()
 		if !ok {
-			// p.err(diag.SynExpectExpression, "expected expression after '='")
-			// todo: попробуем так же посмотреть вокруг, если там пробелы - забираем их тоже
-			p.emitDiagnostic(
-				diag.SynExpectExpression,
-				diag.SevError,
-				tokAssign.Span,
-				"expected expression after '='",
-				func(b *diag.ReportBuilder) {
-					if b == nil {
-						return
-					}
-					fixID := fix.MakeFixID(diag.SynExpectExpression, tokAssign.Span)
-					suggestion := fix.DeleteSpan(
-						"remove '=' to simplify the let binding",
-						tokAssign.Span,
-						"",
-						fix.WithID(fixID),
-						fix.WithKind(diag.FixKindRefactor),
-						fix.WithApplicability(diag.FixApplicabilityAlwaysSafe), // todo подумать безопасно ли это
-					)
-					b.WithFixSuggestion(suggestion)
-					b.WithNote(tokAssign.Span, "remove '=' to simplify the let binding")
-				},
-			)
+			if p.opts.CurrentErrors == beforeErrors {
+				p.emitDiagnostic(
+					diag.SynExpectExpression,
+					diag.SevError,
+					tokAssign.Span,
+					"expected expression after '='",
+					func(b *diag.ReportBuilder) {
+						if b == nil {
+							return
+						}
+						fixID := fix.MakeFixID(diag.SynExpectExpression, tokAssign.Span)
+						suggestion := fix.DeleteSpan(
+							"remove '=' to simplify the let binding",
+							tokAssign.Span,
+							"",
+							fix.WithID(fixID),
+							fix.WithKind(diag.FixKindRefactor),
+							fix.WithApplicability(diag.FixApplicabilityAlwaysSafe),
+						)
+						b.WithFixSuggestion(suggestion)
+						b.WithNote(tokAssign.Span, "remove '=' to simplify the let binding")
+					},
+				)
+			}
 			return LetBinding{}, false
 		}
 	}
