+++ testdata/test_fixes/let_fixes/let_arr.sg
@@ -1,4 +1,4 @@
 let a: int[]; // нет ошибки
 let b: int[ // две ошибки, отсутствие ']' и отсутствие ;
 let c: int[10; // SYN2201 surge fix должен поставить ]
-let d: int[] = [1,];
+let d: int[] = [1,; // добавить скобку
